+++ src/DependencyInjection/Container.php
@@ -4,6 +4,7 @@ namespace Wadify\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
 final class Container
@@ -26,18 +27,33 @@ final class Container
     }
 
     /**
-     * @param string $serviceName
+     * @param string $id
+     * @param array  $addToArgument
      *
      * @return mixed
      */
-    public static function get($serviceName)
+    public static function get($id, array $addToArgument = [])
     {
         if (self::$container === null) {
             self::load();
         }
-        self::$container->compile();
 
-        return self::$container->get($serviceName);
+        if (!empty($addToArgument)) {
+            $definition = self::$container->getDefinition($id);
+            foreach ($addToArgument as $key => $value) {
+                $arguments = $definition->getArguments();
+                foreach ($value as $newKey => $newValue) {
+                    $arguments[$key][$newKey] = $newValue;
+                }
+                $definition->setArguments($arguments);
+            }
+        }
+
+        if (false === self::$container->isFrozen()) {
+            self::$container->compile();
+        }
+
+        return self::$container->get($id);
     }
 
     /**
@@ -46,7 +62,12 @@ final class Container
      */
     public static function set($id, $service)
     {
-        self::load();
+        if (self::$container === null || self::$container->isFrozen()) {
+            self::load();
+        }
+
         self::$container->set($id, $service);
+        $definition = new Definition(get_class($service));
+        self::$container->setDefinition($id, $definition);
     }
 }